requirements.txt
shapely>=2.0.0
pyspark==3.5.2
apache-sedona==1.7.2
pytest==8.3.3
duckdb>=1.3.2
mitmproxy>=11.0.0
pandas>=1.5.0,<3.0.0
plotly>=5.0.0,<6.0.0
pyarrow>=10.0.0
boto3>=1.26.0
